added a lot, new structure and starting on repo's and services

+++ src/main/java/com/realdolmen/shopr/repository/BookFictionRepository.java
@@ -12,7 +12,7 @@ public class BookFictionRepository {
     @PersistenceContext(unitName = "shoprPU")
     private EntityManager entityManager;
 
-    public BookFiction findById(int id) {
+    public BookFiction findById(Long id) {
         return entityManager.find(BookFiction.class, id);
     }
